Moved plotly to dev-dependencies
