+++ Cargo.toml
@@ -16,7 +16,7 @@ documentation = "https://docs.rs/simple_accumulator"
 [dependencies]
 num = "0.4.0"
 rand = "0.8.5"
-plotly = "0.6.0"
+
 
 [dependencies.float_eq]
 version = "1"
@@ -29,4 +29,11 @@ opt-level = 3
 
 [dev-dependencies]
 ordered-float = "3.0.0"
+plotly = "0.6.0"
+
+[lib]
+name = "simple_accumulator"
+
+[[example]]
+name = "error_mean"
